+++ tool/minifier/class/changelog.md
@@ -1,7 +1,9 @@
 ## Change Log
 
 ### 1.1.0 - 2014/11/04
-- Added the ability to minify CSS rules defined in a variable.
+- Added the ability to minify JavaScript scripts defined in a variable that uses the heredoc syntax.
+- Added the ability to minify CSS rules defined in a variable that uses the heredoc syntax.
+- Added the JSMinPlus library.
 
 ### 1.0.1 - 2014/08/14
 - Fixed white spaces.